+++ src/main/java/chess/domain/piece/BlackPawn.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import chess.domain.piece.rule.BlackPawnCatchRule;
 import chess.domain.piece.rule.BlackPawnMoveRule;
+import chess.domain.piece.rule.Rule;
 import chess.domain.position.Position;
 import chess.domain.position.PositionDifference;
 import chess.domain.position.RankDifference;
